Remove WPT and lock step requirement
A new capability has been added to pass on webdriver actions:
commandWebhook

This capability should point to an end point which will receive a POST
with a SAWebdriverAction containing the command name and timestamp of
the action.
