+++ java/client/src/org/openqa/selenium/firefox/FirefoxDriver.java
@@ -37,12 +37,9 @@ import org.openqa.selenium.OutputType;
 import org.openqa.selenium.Platform;
 import org.openqa.selenium.Proxy;
 import org.openqa.selenium.SessionNotCreatedException;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.firefox.internal.NewProfileExtensionConnection;
 import org.openqa.selenium.firefox.internal.ProfilesIni;
-import org.openqa.selenium.html5.LocalStorage;
-import org.openqa.selenium.html5.SessionStorage;
 import org.openqa.selenium.internal.Killable;
 import org.openqa.selenium.internal.Lock;
 import org.openqa.selenium.internal.SocketLock;
@@ -112,7 +109,6 @@ public class FirefoxDriver extends RemoteWebDriver implements Killable {
   public static final String BINARY = "firefox_binary";
   public static final String PROFILE = "firefox_profile";
   public static final String MARIONETTE = "marionette";
-  public static final String PROFILE_DIR = "firefox_profile_dir";
 
   // For now, only enable native events on Windows
   public static final boolean DEFAULT_ENABLE_NATIVE_EVENTS = Platform.getCurrent().is(WINDOWS);
@@ -166,18 +162,6 @@ public class FirefoxDriver extends RemoteWebDriver implements Killable {
           throw new WebDriverException(e);
         }
       }
-    }
-    else {
-      if (desiredCapabilities != null && desiredCapabilities.getCapability(PROFILE_DIR) != null) {
-        raw = desiredCapabilities.getCapability(PROFILE_DIR);
-      }
-      if (requiredCapabilities != null && requiredCapabilities.getCapability(PROFILE_DIR) != null) {
-        raw = requiredCapabilities.getCapability(PROFILE_DIR);
-      }
-      if (raw instanceof String) {
-        profile = new FirefoxProfile(new File((String) raw));
-      }
-
     }
     profile = getProfile(profile);
