+++ java/server/src/com/appdynamics/webhook/BUCK
@@ -1,5 +1,5 @@
 
-java_library(name = "wpt",
+java_library(name = "webhook",
   srcs = glob(["*.java"]),
   deps = [
     '//java/client/src/org/openqa/selenium:selenium',
