+++ java/server/src/org/openqa/selenium/remote/server/BUCK
@@ -49,7 +49,7 @@ java_library(name = 'server',
     '//java/client/src/org/openqa/selenium/remote:remote',
     '//java/client/src/org/openqa/selenium/support/events:events',
     '//java/server/src/org/openqa/selenium/remote/server/log:log',
-    '//java/server/src/com/appdynamics/wpt:wpt',
+    '//java/server/src/com/appdynamics/webhook:webhook',
     '//third_party/java/gson:gson',
     '//third_party/java/guava:guava',
     '//third_party/java/servlet:servlet-api',
