+++ java/server/src/org/openqa/selenium/remote/server/DefaultSession.java
@@ -270,7 +270,7 @@ public class DefaultSession implements Session {
        * @since 7/1/15
        * Copyright 2015 AppDynamics inc. All rights reserved
        */
-      caps.setCapability(CapabilityType.WPT_LOCK_STEP, this.capabilities.is(CapabilityType.WPT_LOCK_STEP));
+      caps.setCapability(CapabilityType.COMMAND_WEBHOOK, this.capabilities.getCapability(CapabilityType.COMMAND_WEBHOOK));
 
       return caps;
     }
